Move boto3_connection.py from `core` to `cloudify_aws`
