+++ Lambda/__init__.py
@@ -24,7 +24,8 @@ from botocore.exceptions import ClientError
 
 from cloudify.exceptions import NonRecoverableError
 from cloudify.decorators import operation
-from core.boto3_connection import connection
+
+from cloudify_aws.boto3_connection import connection
 
 
 @contextmanager
