+++ cloudify_aws/boto3_connection.py
